Add node neighbours
Node neighbours were previously absent, now the VonNeumann cell relation-
ship, where the node belongs to a grid and is connected to the upper,
lower, right and left nodes to him. Obstacles are ignored for these
relationships.

In future changes a relationship with the diagonal nodes will be included
and even a hexagonal relationship, but these are for the future me to
deal with.